Phase-03/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
uvicorn==0.24.0
streamlit==1.28.0
python-dotenv==1.0.0
groq==0.4.1
pydantic==2.5.0
requests==2.31.0